third_party/WebKit/LayoutTests/http/tests/security/suborigins/suborigin-invalid-names-expected.txt

ALERT: I am a secret
CONSOLE ERROR: Error with Suborigin header: Invalid character ''' in suborigin.
ALERT: I am a secret
CONSOLE ERROR: Error with Suborigin header: Invalid character ''' in suborigin.
ALERT: I am a secret
CONSOLE ERROR: Error with Suborigin header: Invalid character ''' in suborigin.
ALERT: I am a secret
CONSOLE ERROR: Error with Suborigin header: Invalid character ''' in suborigin.
ALERT: I am a secret
CONSOLE ERROR: Error with Suborigin header: Invalid character '@' in suborigin.
ALERT: I am a secret
CONSOLE ERROR: Error with Suborigin header: Whitespace is not allowed in suborigin names.
ALERT: I am a secret
